quick fix : 'if' statement with identical branches

+++ src/main/java/tconstruct/modifiers/tools/ModButtertouch.java
@@ -86,10 +86,10 @@ public class ModButtertouch extends ModBoolean {
             if (index == enchant.effectId) {
                 hasEnchant = true;
                 enchantTag.setShort("id", (short) index);
-                enchantTag.setShort("lvl", (byte) level);
+                enchantTag.setShort("lvl", (short) ((byte) level));
             } else {
                 enchantTag.setShort("id", (short) index);
-                enchantTag.setShort("lvl", (byte) lvl);
+                enchantTag.setShort("lvl", (short) ((byte) lvl));
             }
             tags.appendTag(enchantTag);
         }
